+++ src/main/java/roomescape/dao/ReservationRepository.java
@@ -1,4 +1,4 @@
-package roomescape;
+package roomescape.dao;
 
 import java.time.LocalDate;
 import java.time.LocalTime;
@@ -8,6 +8,8 @@ import java.util.Map;
 import org.springframework.jdbc.core.JdbcTemplate;
 import org.springframework.jdbc.core.simple.SimpleJdbcInsert;
 import org.springframework.stereotype.Repository;
+import roomescape.domain.Reservation;
+import roomescape.domain.ReservationTime;
 
 @Repository
 public class ReservationRepository {
