+++ src/main/java/roomescape/dao/ReservationTimeRepository.java
@@ -1,4 +1,4 @@
-package roomescape;
+package roomescape.dao;
 
 import java.time.LocalTime;
 import java.util.HashMap;
@@ -7,6 +7,7 @@ import java.util.Map;
 import org.springframework.jdbc.core.JdbcTemplate;
 import org.springframework.jdbc.core.simple.SimpleJdbcInsert;
 import org.springframework.stereotype.Repository;
+import roomescape.domain.ReservationTime;
 
 @Repository
 public class ReservationTimeRepository {
